requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.16.0
pdfplumber>=0.10.0
httpx>=0.26.0
python-dotenv>=1.0.0
markdown>=3.5.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Payment
stripe>=8.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6

# Settings
pydantic-settings>=2.1.0
email-validator>=2.1.0
